chore: reorganize submodules to follow PEP8 naming convention
