+++ torchopt/__init__.py
@@ -15,50 +15,56 @@
 """TorchOpt: a high-performance optimizer library built upon PyTorch."""
 
 from torchopt._src import (
-  accelerated_op_available,
-  clip,
-  combine,
-  hook,
-  schedule,
-  visual,
+    accelerated_op_available,
+    clip,
+    combine,
+    hook,
+    schedule,
+    visual,
 )
 from torchopt._src.alias import adam, rmsprop, sgd
-from torchopt._src.MetaOptimizer import (
-  MetaAdam,
-  MetaOptimizer,
-  MetaRMSProp,
-  MetaSGD,
+from torchopt._src.optimizer import (
+    Optimizer,
+    SGD,
+    Adam,
+    RMSProp,
+    meta,
 )
-from torchopt._src.Optimizer import SGD, Adam, Optimizer, RMSProp
+from torchopt._src.optimizer.meta import (
+    MetaOptimizer,
+    MetaSGD,
+    MetaAdam,
+    MetaRMSProp,
+)
 from torchopt._src.update import apply_updates
 from torchopt._src.utils import (
-  extract_state_dict,
-  recover_state_dict,
-  stop_gradient,
+    extract_state_dict,
+    recover_state_dict,
+    stop_gradient,
 )
 
 __version__ = "0.4.1"
 
 __all__ = (
-  "accelerated_op_available",
-  "clip",
-  "combine",
-  "hook",
-  "schedule",
-  "visual",
-  "adam",
-  "rmsprop",
-  "sgd",
-  "MetaAdam",
-  "MetaOptimizer",
-  "MetaRMSProp",
-  "MetaSGD",
-  "SGD",
-  "Adam",
-  "Optimizer",
-  "RMSProp",
-  "apply_updates",
-  "extract_state_dict",
-  "recover_state_dict",
-  "stop_gradient",
+    "accelerated_op_available",
+    "clip",
+    "combine",
+    "hook",
+    "schedule",
+    "visual",
+    "adam",
+    "rmsprop",
+    "sgd",
+    "Optimizer",
+    "SGD",
+    "Adam",
+    "RMSProp",
+    "MetaOptimizer",
+    "MetaSGD",
+    "MetaAdam",
+    "MetaRMSProp",
+    "apply_updates",
+    "extract_state_dict",
+    "recover_state_dict",
+    "stop_gradient",
 )
