+++ torchopt/_src/optimizer/__init__.py
@@ -13,4 +13,9 @@
 # limitations under the License.
 # ==============================================================================
 
-from torchopt._src.accelerated_op.adam_op.adam_op import AdamOp
+from torchopt._src.optimizer.base import Optimizer
+from torchopt._src.optimizer.adam import Adam
+from torchopt._src.optimizer.rmsprop import RMSProp
+from torchopt._src.optimizer.sgd import SGD
+
+from torchopt._src.optimizer import meta
