+++ torchopt/_src/optimizer/meta/sgd.py
@@ -1,0 +1,54 @@
+# Copyright 2022 MetaOPT Team. All Rights Reserved.
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#     http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+# ==============================================================================
+
+from typing import Union
+
+import torch.nn as nn
+
+from torchopt._src.alias import sgd
+from torchopt._src.optimizer.meta.base import MetaOptimizer
+from torchopt._src.typing import ScalarOrSchedule
+
+
+class MetaSGD(MetaOptimizer):
+    """A canonical Stochastic Gradient Descent optimizer."""
+
+    def __init__(
+        self,
+        net: nn.Module,
+        lr: ScalarOrSchedule,
+        momentum: Union[float, None] = None,
+        nesterov: bool = False,
+        moment_requires_grad: bool = True
+    ):
+        """The `init` function.
+
+        Args:
+            net (nn.Module):
+                A network whose parameters should be optimized.
+            args:
+                Other arguments see `alias.sgd`, here we set `moment_requires_grad=True`
+                to make tensors like momentum be differentiable.
+        """
+
+        super().__init__(
+            net,
+            sgd(
+                lr=lr,
+                momentum=momentum,
+                nesterov=nesterov,
+                moment_requires_grad=moment_requires_grad
+            )
+        )
